[Silver V] Title: Dogs and Cats, Time: 304 ms, Memory: 23272 KB, Score: 11 point -BaekjoonHub

+++ 백준/Silver/23261. Dogs and Cats/Dogs and Cats.java
@@ -16,7 +16,7 @@ class Main {
         	
         	int N = Integer.parseInt(st.nextToken());
         	int D = Integer.parseInt(st.nextToken());
-        	int C = Integer.parseInt(st.nextToken());
+        	long C = Long.parseLong(st.nextToken());
         	int M = Integer.parseInt(st.nextToken());
         	
         	int dCount = 0;
